+++ v3/lints/cabf_br/lint_dnsname_contains_empty_label_test.go
@@ -1,4 +1,4 @@
-package rfc
+package cabf_br
 
 /*
  * ZLint Copyright 2021 Regents of the University of Michigan
